Refactor: Improve UI and network handling in Seller Profile and related screens
This commit refactors several fragments related to the Seller Profile, including Shop, Shows, Reviews, and introduces a new Clips tab. It focuses on improving UI consistency, handling network errors more gracefully by showing "No Internet" views, and refining data loading states. Additionally, it changes dimension types in `GetMyInventoryResponse` to `Double` and disables user input for the ViewPager in `SellerProfileActivity`.

**Key Changes:**

- **Seller Profile (`SellerProfileActivity.kt`, `activity_seller_profile.xml`):**
    - Disabled user swipe (`isUserInputEnabled = false`) on the `ViewPager2` in `SellerProfileActivity`.
    - Added `NoInternet` and `LoaderView` to `activity_seller_profile.xml` for better loading and error state handling at the activity level.

- **Shop Tab (`ShopFragment.kt`, `fragment_shop.xml`):**
    - Added `NoDataView` (for no products) and `NoInternet` view.
    - Added `LoaderView` for loading states.
    - Implemented `onClick` for `noInternet` view to retry loading products.
    - Updated data loading logic:
        - Clears `productList` only if `page == 1`.
        - Appends new data to `productList`.
        - Shows/hides `noData`, `recycler`, and `noInternet` views based on data and network status.
        - Sets `isLoading` based on `totalPage`.

- **Shows Tab (`SellerShowFragment.kt`, `fragment_seller_show.xml`):**
    - Added `NoDataView` (for no shows, text "No Shows Found") and `NoInternet` view, both wrapped in a `NestedScrollView`.
    - Added `LoaderView` for loading states.
    - Implemented `onClick` for `noInternet` view to retry loading shows.
    - Updated data loading logic:
        - Clears `showList` only if `page == 1`.
        - Appends new data to `showList`.
        - Shows/hides `noData`, `recycler`, and `noInternet` views based on data and network status.
        - Sets `isLoading` based on `totalPage`.

- **Reviews Tab (`ReviewListFragment.kt`, `fragment_review_list.xml`):**
    - Added `NoDataView` (for no reviews, text "No Reviews Found") and `NoInternet` view, both wrapped in a `NestedScrollView`.
    - Added `LoaderView` for loading states.
    - Implemented `onClick` for `noInternet` view to retry loading reviews.
    - Updated data loading logic:
        - Shows/hides `noData`, `recycler`, and `noInternet` views based on data and network status.

- **Clips Tab (New):**
    - Created `ClipsFragment.kt` and `fragment_clips.xml`.
    - `ClipsFragment` displays a grid of clips using `ClipsAdapter`.
    - `fragment_clips.xml` includes a `RecyclerView` (GridLayoutManager with spanCount 2), `NoDataView` ("No Shows Found"), `NoInternet` view, and `LoaderView`, all within a `NestedScrollView`.
    - Created `ClipsAdapter.kt` and `clips_item.xml` for displaying individual clips (currently shows static "Category Name" and avatar).
    - `ClipsFragment` currently shows an empty state by default.

- **`ViewPagerAdapter.kt`:**
    - Modified to include `ClipsFragment` as the fourth tab in the "Shop" (Seller Profile) context, replacing `PurchasesFragment`.

- **Data Model (`GetMyInventoryResponse.kt`):**
    - Changed the type of `width`, `height`, `length`, and `weight` fields from `Int?` to `Double?` to support decimal values.

- **Network & Error Handling Improvements (General):**
    - **`PurchasesFragment.kt`:**
        - Ensured `noInternet` view is hidden on successful data load or when retrying.
        - Corrected loader and swipe refresh visibility management in success and error blocks.
    - **`BidsFragment.kt`:**
        - Ensured `noInternet` view is shown/hidden correctly based on network status and data presence.
        - When offline, `noData` is now hidden if `noInternet` is visible.
    - **`MessagesFragment.kt`:**
        - Minor code formatting and cleanup.

- **Minor UI/Layout:**
    - `fragment_category.xml`: Adjusted `layout_constraintVertical_bias` for `recyclerView`.
    - `WalletViewFragment.kt`: Minor code formatting.

+++ app/src/main/java/io/bidswipe/app/network/response/GetMyInventoryResponse.kt
@@ -51,13 +51,13 @@ data class GetMyInventoryResponse(
         @SerializedName("quantity")
         val quantity: Int?,
         @SerializedName("width")
-        val width: Int?,
+        val width: Double?,
         @SerializedName("height")
-        val height: Int?,
+        val height: Double?,
         @SerializedName("length")
-        val length: Int?,
+        val length: Double?,
         @SerializedName("weight")
-        val weight: Int?,
+        val weight: Double?,
         @SerializedName("mail_class")
         val mailClass: String?,
         @SerializedName("processing_category")
